refactor [4] Issue List 타입스크립트 리팩토링
