+++ FE/tracker/src/components/issueList/issueListHead.tsx
@@ -1,3 +1,12 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import TableCell from "@mui/material/TableCell";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import TableSortLabel from "@mui/material/TableSortLabel";
+import Checkbox from "@mui/material/Checkbox";
+import { visuallyHidden } from "@mui/utils";
+
 import {
     Data,
     HeadCell,
@@ -5,3 +14,96 @@ import {
     EnhancedTableProps,
     EnhancedTableToolbarProps,
 } from "./IssueLIst.types";
+
+const headCells: readonly HeadCell[] = [
+    {
+        id: "name",
+        numeric: false,
+        disablePadding: true,
+        label: "Dessert (100g serving)",
+    },
+    {
+        id: "calories",
+        numeric: true,
+        disablePadding: false,
+        label: "Calories",
+    },
+    {
+        id: "fat",
+        numeric: true,
+        disablePadding: false,
+        label: "Fat (g)",
+    },
+    {
+        id: "carbs",
+        numeric: true,
+        disablePadding: false,
+        label: "Carbs (g)",
+    },
+    {
+        id: "protein",
+        numeric: true,
+        disablePadding: false,
+        label: "Protein (g)",
+    },
+];
+
+function EnhancedTableHead(props: EnhancedTableProps) {
+    const {
+        onSelectAllClick,
+        order,
+        orderBy,
+        numSelected,
+        rowCount,
+        onRequestSort,
+    } = props;
+    const createSortHandler =
+        (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
+            onRequestSort(event, property);
+        };
+
+    return (
+        <TableHead>
+            <TableRow>
+                <TableCell padding="checkbox">
+                    <Checkbox
+                        color="primary"
+                        indeterminate={
+                            numSelected > 0 && numSelected < rowCount
+                        }
+                        checked={rowCount > 0 && numSelected === rowCount}
+                        onChange={onSelectAllClick}
+                        inputProps={{
+                            "aria-label": "select all desserts",
+                        }}
+                    />
+                </TableCell>
+                {headCells.map((headCell) => (
+                    <TableCell
+                        key={headCell.id}
+                        align={headCell.numeric ? "right" : "left"}
+                        padding={headCell.disablePadding ? "none" : "normal"}
+                        sortDirection={orderBy === headCell.id ? order : false}
+                    >
+                        <TableSortLabel
+                            active={orderBy === headCell.id}
+                            direction={orderBy === headCell.id ? order : "asc"}
+                            onClick={createSortHandler(headCell.id)}
+                        >
+                            {headCell.label}
+                            {orderBy === headCell.id ? (
+                                <Box component="span" sx={visuallyHidden}>
+                                    {order === "desc"
+                                        ? "sorted descending"
+                                        : "sorted ascending"}
+                                </Box>
+                            ) : null}
+                        </TableSortLabel>
+                    </TableCell>
+                ))}
+            </TableRow>
+        </TableHead>
+    );
+}
+
+export default EnhancedTableHead;
